Move some end-to-end-tests in a separate file.
